finished functionality to read dxl state and write goal
I decided to skip ros publishers for now, it was making things more
complicated, with its datatypes etc. I know it removes the ability to
interchange with simulation, but that can come later. I want to focus
on playing with the real robot.

I doctested the rad_to_int_range() and vice versa functions and they
succeed.

Signed-off-by: Lasse Simmelsgaard Børresen <fiffigini@gmail.com>